refactor: Move the base class and concrete classes from algorithm.py to the algorithm directory.
